- Fixed MOC DB Access - Corrected testing

+++ trunk/40_implementation/LIMO/LIMO_OrientDB/src/test/java/nl/fontys/sofa/limo/orientdb/mock/MockOrientDBAccess.java
@@ -14,8 +14,11 @@ public class MockOrientDBAccess extends OrientDBAccess {
 
     @Override
     public void closeConnection() {
-        connection.drop();
-        super.closeConnection();
+        if (connection != null) {
+            if (!connection.isClosed())
+                connection.drop();
+            super.closeConnection();
+        }
     }
 
     @Override
